+++ meta_package_manager/tests/test_cli.py
@@ -190,6 +190,37 @@ class TestCLIList(CLITestCase):
         self.assertNotIn("gem", result.output)
 
 
+class TestCLISearch(CLITestCase):
+
+    def test_main_help(self):
+        result = self.invoke('search', '--help')
+        self.assertEqual(result.exit_code, 0)
+        self.assertIn("--help", result.output)
+
+    def test_verbosity(self):
+        result = self.invoke('--verbosity', 'DEBUG', 'search', 'abc')
+        self.assertEqual(result.exit_code, 0)
+        self.assertIn("debug:", result.output)
+
+        result = self.invoke('--verbosity', 'INFO', 'search', 'abc')
+        self.assertEqual(result.exit_code, 0)
+        self.assertNotIn("debug:", result.output)
+
+    def test_simple_call(self):
+        result = self.invoke('search', 'abc')
+        self.assertEqual(result.exit_code, 0)
+
+    def test_sub_manager_scope(self):
+        result = self.invoke('--manager', 'npm', 'search', 'abc')
+        self.assertEqual(result.exit_code, 0)
+        self.assertIn("npm", result.output)
+        self.assertNotIn("apm", result.output)
+        self.assertNotIn("brew", result.output)
+        self.assertNotIn("pip2", result.output)
+        self.assertNotIn("pip3", result.output)
+        self.assertNotIn("gem", result.output)
+
+
 class TestCLIOutdated(CLITestCase):
 
     def test_main_help(self):